+++ riptide-spring-boot-starter/src/main/java/org/zalando/riptide/spring/Registry.java
@@ -13,7 +13,6 @@ import org.springframework.beans.factory.support.ManagedList;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 import static com.google.common.base.CaseFormat.LOWER_CAMEL;
