+++ riptide-spring-boot-starter/src/main/java/org/zalando/riptide/spring/RiptideTestAutoConfiguration.java
@@ -1,0 +1,16 @@
+package org.zalando.riptide.spring;
+
+import org.springframework.boot.autoconfigure.AutoConfigureBefore;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@AutoConfigureBefore(RiptideAutoConfiguration.class)
+public class RiptideTestAutoConfiguration {
+
+    @Bean
+    public static RiptidePostProcessor restClientTestPostProcessor() {
+        return new RiptidePostProcessor(TestRiptideRegistrar::new);
+    }
+
+}
