+++ riptide-spring-boot-starter/src/test/java/org/zalando/riptide/spring/testing/RiptideClientTestTest.java
@@ -3,12 +3,14 @@ package org.zalando.riptide.spring.testing;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.client.MockRestServiceServer;
+import org.zalando.riptide.spring.RiptideClientTest;
 
 import static org.springframework.test.web.client.match.MockRestRequestMatchers.requestTo;
 import static org.springframework.test.web.client.response.MockRestResponseCreators.withSuccess;
@@ -29,6 +31,7 @@ public class RiptideClientTestTest {
     private TestService client;
 
     @Autowired
+    @Qualifier("example")
     private MockRestServiceServer server;
 
     @Test
@@ -38,10 +41,4 @@ public class RiptideClientTestTest {
         server.verify();
     }
 
-    @Test
-    public void shouldAutowireMockedRestTemplate() throws Exception {
-        server.expect(requestTo("https://example.com/foo/bar")).andRespond(withSuccess());
-        client.callViaRestTemplate();
-        server.verify();
-    }
 }
