+++ riptide-spring-boot-starter/src/test/java/org/zalando/riptide/spring/testing/TestService.java
@@ -2,10 +2,6 @@ package org.zalando.riptide.spring.testing;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 import org.zalando.riptide.Http;
@@ -31,10 +27,4 @@ public class TestService {
         http.get("/bar").dispatch(status(), on(OK).call(pass())).join();
     }
 
-    void callViaRestTemplate() {
-        final ResponseEntity<Void> response = restTemplate.exchange("/bar", HttpMethod.GET, HttpEntity.EMPTY, Void.class);
-        if (response.getStatusCode() != HttpStatus.OK) {
-            throw new IllegalStateException("Unexpected response");
-        }
-    }
 }
