+++ riptide-spring-boot-starter/src/test/resources/application-testing.yml
@@ -10,3 +10,5 @@ riptide:
         max-size: 100
         keep-alive: 5 minutes
         queue-size: 10
+    other:
+      base-url: http://example.com
